+++ anyblok_wms_base/core/operation/arrival.py
@@ -24,8 +24,9 @@ Operation = Declarations.Model.Wms.Operation
 class Arrival(Operation):
     """Operation to describe physical arrival of goods in some location.
 
-    Arrivals store data about the expected or arrived Goods: properties, code…
-    These are copied over to the corresponding Goods records in all
+    Arrivals store data about the expected or arrived physical objects:
+    properties, code…
+    These are copied over to the corresponding PhysObj records in all
     cases and stay inert after the fact.
 
     In case the Arrival state is ``planned``,
@@ -36,7 +37,7 @@ class Arrival(Operation):
     - maybe the application won't use the ``planned`` state at all, and
       will only create Arrival after checking them,
     - maybe the application will inspect the Arrival properties, compare them
-      to reality, update them on the created Goods and cancel downstream
+      to reality, update them on the created PhysObj and cancel downstream
       operations if needed, before calling :meth:`execute`.
 
     TODO maybe provide higher level facilities for validation scenarios.
@@ -48,28 +49,28 @@ class Arrival(Operation):
                  autoincrement=False,
                  foreign_key=Operation.use('id').options(ondelete='cascade'))
     """Primary key."""
-    goods_type = Many2One(model='Model.Wms.Goods.Type')
-    """Expected :class:`Goods Type
+    goods_type = Many2One(model='Model.Wms.PhysObj.Type')
+    """Expected :class:`PhysObj Type
     <anyblok_wms_base.core.goods.Type>`.
     """
-    goods_properties = Jsonb(label="Properties of arrived Goods")
+    goods_properties = Jsonb(label="Properties of arrived PhysObj")
     """Expected :class:`Properties
     <anyblok_wms_base.core.goods.Properties>`.
 
-    They are copied over to the newly created :class:`Goods
-    <anyblok_wms_base.core.goods.Goods>` as soon as the Arrival
+    They are copied over to the newly created :class:`PhysObj
+    <anyblok_wms_base.core.goods.PhysObj>` as soon as the Arrival
     is planned, and aren't updated by :meth:`execute`. Matching them with
     reality is the concern of separate validation processes, and this
     field can serve for later assessments after the fact.
     """
-    goods_code = Text(label="Code to set on arrived Goods")
-    """Expected :attr:`Goods code
-    <anyblok_wms_base.core.goods.Goods.code>`.
+    goods_code = Text(label="Code to set on arrived PhysObj")
+    """Expected :attr:`PhysObj code
+    <anyblok_wms_base.core.goods.PhysObj.code>`.
 
     Can be ``None`` in case the arrival process issues the code only
     at the time of actual arrival.
     """
-    location = Many2One(model='Model.Wms.Goods')
+    location = Many2One(model='Model.Wms.PhysObj')
     """Will be the location of the initial Avatar."""
 
     inputs_number = 0
@@ -91,17 +92,17 @@ class Arrival(Operation):
                 offender=location)
 
     def after_insert(self):
-        Goods = self.registry.Wms.Goods
+        PhysObj = self.registry.Wms.PhysObj
         self_props = self.goods_properties
         if self_props is None:
             props = None
         else:
-            props = Goods.Properties.create(**self_props)
+            props = PhysObj.Properties.create(**self_props)
 
-        goods = Goods.insert(type=self.goods_type,
-                             properties=props,
-                             code=self.goods_code)
-        Goods.Avatar.insert(
+        goods = PhysObj.insert(type=self.goods_type,
+                               properties=props,
+                               code=self.goods_code)
+        PhysObj.Avatar.insert(
             goods=goods,
             location=self.location,
             reason=self,
@@ -110,6 +111,6 @@ class Arrival(Operation):
         )
 
     def execute_planned(self):
-        Avatar = self.registry.Wms.Goods.Avatar
+        Avatar = self.registry.Wms.PhysObj.Avatar
         Avatar.query().filter(Avatar.reason == self).one().update(
             state='present', dt_from=self.dt_execution)