+++ anyblok_wms_base/core/tests/__init__.py
@@ -15,12 +15,12 @@ class TestCore(BlokTestCase):
 
     def test_minimal(self):
         Wms = self.registry.Wms
-        goods_type = Wms.Goods.Type.insert(label="My good type",
-                                           code='MyGT')
+        goods_type = Wms.PhysObj.Type.insert(label="My good type",
+                                             code='MyGT')
         self.assertEqual(goods_type.label, "My good type")
 
-        location_type = Wms.Goods.Type.insert(code="LOC")
-        loc = Wms.Goods.insert(code="Root", type=location_type)
+        location_type = Wms.PhysObj.Type.insert(code="LOC")
+        loc = Wms.PhysObj.insert(code="Root", type=location_type)
         arrival = Wms.Operation.Arrival.insert(goods_type=goods_type,
                                                location=loc,
                                                dt_execution=datetime.now(),