+++ anyblok_wms_base/quantity/operation/tests/test_splitter.py
@@ -6,14 +6,14 @@
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
-from anyblok_wms_base.testing import WmsTestCaseWithGoods
+from anyblok_wms_base.testing import WmsTestCaseWithPhysObj
 from anyblok_wms_base.exceptions import (
     OperationQuantityError,
     OperationMissingQuantityError,
 )
 
 
-class TestSplitterOperation(WmsTestCaseWithGoods):
+class TestSplitterOperation(WmsTestCaseWithPhysObj):
     """Test the WmsSingleGoodOperation mixin
 
     In these test cases, Operation.Move is considered the canonical example of
@@ -125,4 +125,4 @@ class TestSplitterOperation(WmsTestCaseWithGoods):
         str(move)
 
 
-del WmsTestCaseWithGoods
+del WmsTestCaseWithPhysObj