+++ doc/CHANGES.rst
@@ -9,8 +9,16 @@
 Release history
 ===============
 
-0.7.0 (unreleased)
+0.8.0 (unreleased)
 ~~~~~~~~~~~~~~~~~~
+
+* Merge of the ``Wms.Goods`` and ``Wms.Location`` Models into a single
+  one: ``Wms.PhysObj``. This impacts all existing code bases.
+* Inventory Operations: Apparition, Disparition and Teleportation
+* Enrichment of Properties API
+
+0.7.0
+~~~~~
 * Moved the quantity field of Goods from wms-core to the new
   optional wms-quantity Blok.